# specs/akep/v0.1/openapi.yaml
openapi: 3.1.2
...
info:
  title: Agent Knowledge Exchange Protocol
  version: 0.1.0
  summary: Experimental HTTP binding for persistent, traceable Agent knowledge.
  description: |
    AKEP v0.1 Reader, Contributor, Evaluator, Curator, Publisher, and Federation operations.
    The capability document at `/.well-known/akep` supplies the actual base URL,
    supported profiles, authorization metadata, limits, and extensions.
  license:
    name: Project license pending
    identifier: LicenseRef-Pending
servers:
  - url: https://knowledge.example/akep/0.1
    description: Illustrative server; replace with discovery baseUrl
security:
  - oauth2: []
...
paths:
  /.well-known/akep:
    get:
      tags: [Reader]
      operationId: discoverAKEP
      summary: Discover protocol versions, profiles, authorization, and limits
      security: []
      servers:
        - url: https://knowledge.example
          description: Node origin rather than the discovered AKEP base URL
      responses:
        '200':
          description: Short-lived capability document
          headers:
            AKEP-Version:
              $ref: '#/components/headers/AKEPVersion'
            Cache-Control:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: ./schemas/capability.schema.json
        '404':
          $ref: '#/components/responses/Problem'
...
components:
...
  headers:
    AKEPVersion:
      description: Negotiated AKEP protocol version
      schema:
        const: '0.1'
...
  responses:
...
    Problem:
      description: RFC 9457 problem details
      headers:
        AKEP-Version:
          $ref: '#/components/headers/AKEPVersion'
        Cache-Control:
          schema:
            const: 'private, no-store'
      content:
        application/problem+json:
          schema:
            $ref: ./schemas/problem.schema.json
  securitySchemes:
    oauth2:
      type: oauth2
      description: |
        Illustrative OAuth metadata. Clients must use the authorization server
        discovered from RFC 9728 protected resource metadata.
      flows:
        clientCredentials:
          tokenUrl: https://identity.example/oauth2/token
          scopes:
            akep:read: Read authorized immutable revisions
            akep:query: Query authorized knowledge
            akep:contribute: Submit candidate contributions
            akep:feedback: Submit usage evidence
            akep:review: Review and decide candidate contributions
            akep:publish: Publish or deprecate verified contributions
            akep:incident: Apply emergency revoke to a verified revoke contribution
            akep:erase: Apply privacy/legal erasure to a verified erase contribution
            akep:federate: Exchange signed federation changes